MVS: move shape loading helpers to helpers/load-shape.ts
load-helpers.ts was getting too big. Move the private `shape` node code
(ShapeCustomProps, shapeColor, shapeRepresentationProps,
matrixFromTransformNode, shapeTransforms) into its own file.

No behavior change; only the ParamDefinition import moves with it, as the
shape block was its only consumer in load-helpers.ts.
